Tests/LibWeb/Text/expected/UIEvents/KeyEvent-alt.txt

key="l" code=KeyL altKey=true
